Significantly update document.c to "understand" newlines in the output. This is required by the nroff front-end to manage span-level tags (parse_inline) that call newline-sensitive macros, e.g., the new pdfmark.
Using this understanding, expand nroff to allow the GNU pdfmark extension used
(right now) for pdfhref links in the output.

This is a preliminary effort and may have some flaws.  The pdfmark macros
have the flaws of ignoring leading and trailing text that abuts the macro
(should use, e.g., the -A flag to pdfhref).  This will come later.