Add some PHP extra markdown attributes to links.
This adds PHP extra markdown attributes to links and images.  At
this time, this is limited to IDs, classes, and width and height
style attributes.  Others are ignored.

This will work with links, images, and references.

Output is limited to -Thtml and -Tlatex.

Existing support for extended image attributes has been removed in
favour of this more robust parsing.