Fix troff compatibility & allow overriding horizontal rule in -Tms.
This commit does three things for -Tms output:

Fixes the generation of a horizontal rule for compatibility with
(Heirloom) troff. This only removes the calculation of indent, which is
removed with the previous LP call anyway.

Adds a single line of vertical space on either side of a horizontal line.

Allows the user to define a HR macro, which, if defined, will be used in
place of the generated horizontal rule. This allows for fancy literary
styles of typesetting.