Simplify groff's .pdfhref handling.
This vastly simplifies the nroff.c internals by no longer needing
to offset from content.  The original offset code was to handle
situations like:

 foo[link](www.bsd.lv)bar

In this, the "foo" and "bar" needed to be repurposed as appended to
or prefixing the link, as otherwise .pdfhref would pad the output
in spaces.

Instead of doing this, emit "foo" as normal, but have a special
scoping BSCOPE_PDFHREF that detects the prior node and emits the
"\c" after it.  Similarly, if the next node is text without a space,
append the "\c".

There's still work that needs to be done, but this is a start.