Use pandoc's section IDs for HTML output.
Prior to this, lowdown was using section IDs created by URL-ifying
the section contents. Pandoc, on the other hand, uses a simpler
method of just removing everything but alnum, some specials, and
replacing spaces with hyphens.  Use this, and put it in a utility
function for use by other frontends.