Have definition lists use an "empty" IP.
This simplifies how definition lists are handled: instead of using
IP with its width specified by the first line contents, use IP and
a default offset with the data being on the next line.

The problem is with pdfhref, of course.  I couldn't find any way
of invoking:

 .IP "\c
 .pdfhref ...

Where the link wasn't being indented.  In the end, this is much
simpler, and having definition title and data on a different line
are stylistically fine.