Fully fix -Tms and -Tman headers.
This covers all basis to have -Tms and -Tman headers, using pandoc
as a reference point.  In the current way, first the code is simplified
by using content after the SH or SS or otherwise section header as the
section contents instead of putting it all on the same line.

This makes embedded links possible since they don't break a line,
so multiple lines with macros are handled.

Second, use pdfhref directly instead of the undocumented XN to have
sections in the PDF contents and for links (yet to be done).  Only
the text content of the containing macros is used.  This makes the
links like pandoc and -Thtml, and also makes the PDF sections be
readable.