Bugfix: "w" character in extended header attrs.
Earlier, the algorithm to verify trailing extended attributes in a
header simply looked at the last node.  However, lowdown will break
apart text strings arbitrarily (e.g., if the "w" is encountered, it
will break into the auto-link code until it sees that's not, but the
output is multiple consecutive text nodes).

This consolidates consecutive text underneath the header.