Allow to selectively install binaries, static or shared lib
This is achieved by splitting the install target into four main ones:

* install: installs the binaries and documentation
* install_lib_common: installs pkg config files, headers
* install_shared: installs the shared object
* install_static: installs the static archive

The equivalent of the install target before this commit would be
install install_shared install_static.

The main benefit of this change is that packagers can choose which of
the main components to install which is especially helpful when you
don't want to distribute a static version of the library. One
imperfection of the approach taken is that the library man pages will
be installed together with the executable.