Handle rolling footnote digits in -tterm.
Prior to this, if the footnotes increased in digits during a table
(e.g., 9 to 10), the -tterm output might crash.  This was because
footnotes were being zeroed prior to the table column scanning part,
then restored.  However, this would mean they started at 1 instead
of (say) 9, so the real run would have an additional character in
the column width that raised an assertion.

Sneaky, sneaky.