"Elegantly" handle totally-different diffs.
In (apparently) all prior versions of the diff tool, a fully
different pair of trees would cause a crash because the root nodes
never had a chance to match.  First, fix this by manually ensuring
that the root nodes match.

Next, make sure that LOWDOWN_DOC_HEADER also matches, because its
place in the tree is fixed.  Without this change, there would be
diff trees in standalone mode with doubly-printed <head> nodes.