Don't propogate footnote paragraph status.
When creating footnotes in -tms, restore the post-paragraph status
(i.e., whether the subsequent paragraph should be indented or not)
to what the original context specified.

This is because footnotes might have paragraphs that alter the
post-paragraph state, and these footnotes will be delayed in their
printing and not actually emit the real output, which is just a
reference.