Make CRLF safe.
If a DOS-style CRLF is detected on the first line, strip out all
CR characters from the input.

This makes the parsing sequence itself easier, as tests for newlines
need only look for \n and not \r\n.

Prior to this, the parser wasn't 100% CRLF safe.

While here, document that the text in Markdown documents is UTF8
with either UNIX or DOS eolns.