Initial addition of pandoc metadata.
This is the initial support for pandoc's metadata alongside multi-
markdown.  Metadata is passed unmolested into the rendering tools with
the exception of the author block, which has a syntax in pandoc
(semicolon separation) not shared by MMD.  For this, re-write the
metadata in these ways:

 1. compress multiple spaces into one to prevent the front-end from
    interpreting MMD multi-spacing as multiple authors;
 2. convert a newline into a space, which will cause multi-lines to have
    two spaces, which for authors will trigger the multi-author mode
    just like in MMD; and
 3. convert a semicolon into a space.

Further commits are required to make this work without issue with the
expectations of pandoc writers (e.g., a title block within the <body>
of an HTML document) and to catch corner cases.

References #117