Restrict what's considered a man title.
Require "title(section)" to parse a mantitle from the title of a
pandoc metadata line.  The only limit put on the section is that
it must begin with a digit or "n" and be followed by only alpha
then a close paren.  This covers all regular UNIX sections and
the special snowflake tcl/tk ("n").