Stop using "Untitled article" as a default.
Untitled article, besides being language-centric, is also just
wrong for any media that's not an article.

Instead, use an empty string and double-check which outputs
actually require titles and in what circumstances.  For example,
LaTeX doesn't need a title unless an author or date is specified.

Update all regression tests and manpages to this effect.