Attempt to solve the roff constant-width problem.
This has been many-times discussed online, e.g.:

https://github.com/jgm/pandoc/issues/9020

The issue at heart is that the "C" font for roff (mandoc, groff, etc.)
is non-standard across output devices.  Prior to this change, lowdown
was emitting the "C" font for constant-width fonts (those `in
backticks`).  These were being discarded by mandoc and groff when on the
terminal, where "C" is unnecessary because everything is "C".  Mandoc
warned about it, but groff just silently let it go by until a recent
version displayed a warning message.

This introduces some changes.  First, "C" is retained by default for
-tms, although output as "CR" instead of "C".  The rationale is that
-ms is mostly used for formatting PDF/PS, which have the "C" font.
Second, "C" is no longer used for -tman: instead, a bold font is used.

However, the user can override this behaviour.  (This will be documented
in subsequent checkins.)  The --nroff-code-font (not yet finalised)
accepts a 4-tuple describing the constant width fonts, specifically
"regular,bold,italic,bold-italic".  This lets the user override the font
that will be coded into the output.  Three aliases exist: "none",
"bold", and "code".  "None" does not render a constant-width font at
all, and simply uses the regular font.  "Bold" uses a bold font
(obviously if it's a **`situation like this`**, the bold is kept as-is).
Finally, "code" uses the constant-width variants.

References #123