Implement pandoc-style superscripts.
Prior to this, superscripts followed the traditional Markdown syntax
(which is also MMD).  Most users, however, use pandoc, which is pretty
bad in comparison: it has convoluted white-space rules and doesn't allow
nested superscripts.

Since these are incompatible, switch to using the pandoc-style ones by
default, and add an option for using the traditional "short" ones.