Add GFM-style superscripts and subscripts.
This is an extensive change that defaults to using GFM style
superscripts, i.e. a^b^ instead of just a^b or a^(b).  This is the
default form for pandoc as well.  This form can be overridden by the new
LOWDOWN_SUPER_SHORT flag and corresponding --parse-super-short option.

While here, allow for subscripts as well.  Subscript syntax overlaps
with ~~strike-thrus~~, which are enabled by default; however,
strike-thrus start with two tildes, while subscripts use only one.