Parse markdown found within HTML blocks
This changeset updates the parsing behavior to pass along the content
between opening and closing tags of a LOWDOWN_BLOCKHTML segment for
additional markdown parsing. The strict tag matching behavior has been
removed, as it lacked the context needed to operate correctly when there
may be nested tags (such as nested divs).

Consider a content segment such as:

```
<div class="container">
<div class="container column-1">

And some text
</div>
<div class="container column-2">

With more text
</div>
</div>
```
With these changes, an outer LOWDOWN_BLOCKHTML node will be the parent
of all the content within the "container" div. The closing tag for each
block is included as the final child as a LOWDOWN_RAW_HTML node.