Don't output a newline after `.SH`
On macOS, `makewhatis` fails to parse `NAME` headings in man pages if
there's a newline between `.SH` and `NAME`. This is silly, but I can't
exactly get them to update it:

```
$ /usr/libexec/makewhatis -v -o /tmp/whatis outputs/out/share/man | head
man directory outputs/out/share/man
  outputs/out/share/man/man5
        reading outputs/out/share/man/man5/nix.conf.5
        ignoring junk description ""
        reading outputs/out/share/man/man5/nix-profiles.5
        ignoring junk description ""
```

Through a circuitous sequence of events, this results in shell
completions for the `man` command being broken in Fish for man pages
generated with Lowdown:
https://git.lix.systems/lix-project/lix/issues/515

This patch, then, replaces the trailing newline after blocks with a
trailing space for `.SH` blocks which only contain normal text in
`nroff` output.

The restriction that `.SH` blocks only contain normal text is because
(for example) this doesn't render correctly:

```
.SH a \fIhttps://foo.com\fR now \fBworld\fR <\fIhttps://bar.com\fR> b
```