Use UR/UE and MT/ME for -man links.
Re-write how links are handled in -man and -ms.  The existing -ms
behaviour is preserved, while -man is now using the UR/UE and MT/ME
macros by default (unless --nroff-no-groff is used).

There are some exceptions due to malformatting by groff or mandoc:
these macros are not used in headers or definition list titles because
the "next-line" concept is not properly handled across all formatters.

The MT/ME macros are used as well, with stripping of "mailto:" where
applicable.  The same is accorded to -ms usage of .pdfhref.

Add lots of regression tests for this behaviour.