Rename "--nroff-xxx" arguments as "--roff-xxx".
This is a change long in coming.  Historically, the use of "nroff"
was because -tman was the output choice, and -tman is mostly used
with nroff for formatting to a terminal.

The use of "roff" is because both -tms and -tman are roff(7)
macros processed by nroff.  So name the argument in favour of the
language, not the tool.

The old prefix is retained and will not go away.