Add --template debugging.
This allows **-s -ttree** to see the template AST, which eases
debugging of complex templates.  It introduce an API change of
the tree interface needing to touch the lowdown options.

While here, considerably clean up the main manpages, introducing
a new section for metadata and cleaning up the standalone part.
This cleanup is due to the -s behaviour changes w/r/t the new
template functionality.