Specify padding/margin for -tterm.
This adds initial support (which will be refactored in subsequent
commits) for specifying both --term-hmargin and --term-hpadding.
The former will move the content to the right; the latter will pad
into the content width toward the right.

In doing so, make the default body padding to be zero, which allows
the users to control finely for horizontal indent.

Fixes #129