Simplify and cohere -tterm images and links.
First, this prints image links similarly to pandoc, and brings
together the formatting between images and regular links.

While here, tidy up term.h to revert infix strings back to just
being strings, as the lenghts are not important.

Lastly, split the "hrule" creation to its own function, as the
behaviour of an hrule is special.