Fix off-by-one table header row glyph in -tterm
When drawing the header row line, one glyph is written per width plus
padding at the end).  Prior to this, the padding in non-leading columns
was not being printed, resulting in a shorter and shorter line with each
column.

Fix this by drawing the additional glyph; and while here, don't draw an
additional row glyph where the final "inter" mark would be, which
extends the line by one.

Document the algorithm in the code too and fix the existing regression
test, adding a new test whilst there.

Fixes #153