Have `make regress` error out on any failures
Prior to this, `make regress` would report errors, but not act on them.
Only `make valgrind` would properly error out, and that only for memory
errors and not underlying regressions.

This changes each `diff` call to catch a non-zero diff and accumulate
it, reporting any errors at the end and propogating it as a failed
`make` rule.  Note: this only catches errors in expected difference, so
if lowdown fails but outputs the "correct" output, the error is not
caught.

This also has `make valgrind` record any underlying `make regress` error
and propogate it as well.

This builds on a patch introduced by #157, thank you!

References cfac04efbca847da44ca5570ab8310a2b84eb514
References #157