Allow for shared-object linking
If `LINK_METHOD` is set in `configure`, recognise it and let it
influence how `lowdown` is linked to its library.  If `LINK_METHOD` is
set to "shared", link to the shared library; otherwise (or if unset),
default to the static library.

This uses the new `./configure` semantics, where extra key-value pairs
are passed directly into `Makefile.configure`.

References #157