Add compats for shared build
When building with the shared library (LINK_METHOD=shared), the
compats.o file needs to be included when linking.  This is because the
default visibility prevents the main.o from seeing the compatibility
functions (e.g., reallocarray) during linking.

This also adds regression tests in the CI for this specific situation.

References #166