Inhibit link for -tterm images in links
Special cases for an image within a link with terminal output.  The
premise is that the outer link must take precedence over the inner image
if its link text is not being printed.

In normal operation, the printing of the link for both inner image and
link give two places for the OSC8 link, so it's fine.

However,wWhen link URLs are inhibited with **--term-no-link** or
similar, this changes: the outer link would previously be lost because
it had nothing to decorate, as the inner link was decorated for its
link.  The same would apply with relative link inhibition when the
parent link is relative.

With this fix, don't have the inner image decorate if those conditions
arise.  The effect is that outer links will "cover" the inner image.

References #165