Tweak OSC8-decorated -tterm link images
Prior to this, images in links weren't handled well with both -tterm and
--term-no-links or equivalent: since the link has no text but the image,
only the image had OSC8 link decoration.  This violates least-surprise,
as one would expect the link to be dominant.

Now, in the situation of a link without text and only an image (or
multiple images), have the unicode icon and (if shown) image URL link to
the image URL, while the image text links to the surrounding link.

The only corner case now is links with images without alt text.  But we
should all be writing alt text, right?

Fixes #165