Inhibit links properly in -tms and -tman
First, inherit the -tterm concept of inhibiting relative links.  This is
useful for -tman and -tms because relative links don't really make sense
in these documents.  This is specified as LOWDOWN_NROFF_NORELLINK.

Next, generalise how links are inhibited to be similar to -tterm.  If
links will be inhibited with LOWDOWN_NROFF_NOLINK or a relative link
with LOWDOWN_NROFF_NORELLINK, behave as if in traditional mode: output
the link text in bold and that's all.  If, however, these are specified
and there is no link text, then output the link URL in italics.

This changes the existing behaviour of --roff-no-links, which is stated
to only apply when --roff-traditional is specified: now the flag will
apply even if not specified.  This means a major number bump will be
required.

References #170