Clarify documentation for link handling
Inhibiting link URLs is difficult across -tms and -tman.  For the
latter, it mostly works, since the non-traditional `pdfhref` macro will
render a clickable link, where the content can be styled in the same way
as for -tterm and such.  Thus, things like --roff-no-links and
--roff-no-rellinks make sense.

For -tman, it's much more complicated because the non-traditional `UR`
macro will always show a full link alongside the content.  This makes
the concept of a short-link redundant, and makes it impossible to hide
the link content itself.  Thus, have the no-links and no-rellinks no
longer use `UR`, and instead show in traditional mode, which can then
handle the various requests of whether to show links or relative links
alongside content.

Compress this tangly logic into the language of "When possible, don't
show..." in the manpage.